backend/pyproject.toml
[project]
name = "backend-sistema-ouvidoria"
version = "0.1.0"
description = ""
authors = [
    {name = "bruno-gabriel-muniz",email = "brunomunizlourenco@gmail.com"}
]
license = {text = "MIT"}
readme = "/README.md"
requires-python = "<4.0, >=3.13"
dependencies = [
    "fastapi[standard] (>=0.121.2,<0.122.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "validate-docbr (>=1.11.1,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.14.5,<0.15.0)",
    "testcontainers (>=4.13.3,<5.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "ipdb (>=0.13.13,<0.14.0)",
    "faker (>=38.2.0,<39.0.0)",
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev src/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=./src -vv'
post_test = 'coverage html'
